advance: automatically advance when using unsafe_substring
unsafe_substring previously did not advance the position of the input,
requiring a separate call to advance with the length of the string in
order to proceed. The code did this to allow for flexibility in the use
of unsafe_substring. This flexibility, however, was not being used and
came at the cost of having to allocate at least two additional parsers.
The code for unsafe_substring now advances the input position,
eliminating the intermediate parser and as a result reducing allocations
substantially.