json: rework string parsing
The code for parsing strings was inefficient on several grounds. First,
it potentially allocated many intermediate strings, only to concatenate
them before returing the result to the user. Second, it used the
backtracking choice combinator, growing the failure continuation
unnecessarily.

The new implementation does not suffer these drawbacks. It parses a
string one character at a time using the skip_while parser and building
up the result in a Buffer.t. All backtracking is eliminated, the string
is parsed using a single loop (assuming that the string does not cross
input chunk boundaries), and no intermate strings are allocated. The
Buffer.t is also reused throughout the course of a parse, no memory is
now needlessly freed and then reallocated.