char: reduce allocations from use of char parsers
The char parsers as implemented previously required the use of binds,
leading to unnecessary allocations of intermediate parser and therefore
closures. With this patch the new char parsers (with the exception of
peek_char) are built from a single _char primitive parser. This parser
requires that there is at least one more character in the input, and
fails otherwise. Its function argument is applied to the character and
and transforms it to an optional value. If the function returns `Some v`
then parser succeeds with value `v`, and fails otherwise.