at_end: remove take_rest
take_rest as implemented was a very memory-inefficient parser. It forced
all the remaining input to be in memory and then made a copy of it. It's
better to implement this functionality outside of the parser entirely:
run the parser that would have preceded it until it succeeeds, and then
read from the input source until an eof is reached.