positions: consistently return success and failure position
The Fail state for the Unbuffered interface previously did not return
indicate where in the input the parse failed. While this is not the most
useful information for parsers that make use of extensive backtracking,
it can help in some cases.

As part of the change, move the position information for both the
Buffered and Unbuffered state types to be the first value contained in
the variant.

Addresses #69