benchmark-reorg: split command into subcommands, add more json
It was inconvenient to run quick benchmarks before due to the fact that
there was no way to select a subset of benchmarks to run. This is fixed
by splitting the benchmarks into subcommands that only run benchmarks
for a specific parser.

The number of json input files has been expanded. These files were taken
from the attoparsec repository on GitHub. An ACKNOWLEDGEMENTS file was
added to note this.

Added a benchmark to see what the relative costs of float_of_string and
int_of_string are.