count_while-optimize: eliminate advance from count_while calls
The `count_while f >>= advance` pattern was a signifcant source of
parser allocations. Elminiate that pattern by passing a `with_buffer`
function to count_while, which will be passed a view into the input
that count_while will consume. The result of `with_buffer` is now the
result of the parser.