input-reorg: make usage of Input module clearer
The names of functions from the input module were pretty bad. This
commit fixes that and in the process expands functionality of the
unbuffered interface. Rather than requiring a bigstring as an input, the
client can specify a range of bytes to use as input within the
bigstring, specified by the `off` and `len` arguments of the `continue`
field of the `partial` types.