bad-check: remove redundant check that hid a bug
This reverses the fix introduced by #104, which actually hid a bug in
the old `Input.uncommitted_bytes`. That function  using the buffer
length rather than the input length to compute the number of uncommitted
bytes. As part of the refactoring performed in #123, this bug was fixed.
So despite the removal of the fix in #104, the test introduced in the
same pull request continues to pass.