count_while-bug: fix bug in count_while and couht_while1
On a failed prompt call, the position was not being reset to the
position at the start of the count_while or count_while1, rather than
taking into account the consumed input. In practice this would be
difficult to detect as the end of input was reached, except when
checking if the next parser was checking if it was the end of the input!

This commit fixes the bug and adds regression tests.