ints: initial implementation
Add int parsers that take an int as an argument and esure that the
corresponding number of bytes match the beginning of the input.

These parsers are useful for checking protocol constants, and should be
more efficient than the newly-named `any_intN` parsers. These previous
parsers required a bind in order to perform this check, whereas the new
parsers do not.