fix-choice-commit: check parser_committed_bytes
The failure contination in choice was previously checking
`client_committed_bytes` to determine if it was OK to backtrack to the
previous position to try the next branch of the choice. The code should
have been checking `parser_committed_bytes` instead, which is the value
that's actually updated by the `commit` parser.

Added a test to catch this in the future.

Closes #144.