add `consumed` combinator
From the implementation, it is a little strange to call this a
combinator since it manually constructs a parser. This helper is useful
for running a complex series of parsers whose purpose is to "validate" a
substring of input without actually creating a concrete type out of it.
The motivating example for this is parsing numbers in JSON -- you want
to make sure the string conforms to the spec, but you don't actually
want to pull apart the individually parsed pieces just to reconstruct
them into the same string.