Fix `Angstrom.advance` to fail if the remaining input is less than n
The `advance` docstring states that `advance` will fail if the remaining
input is less than `n`. However, it seems like this is not actually the
case in practice.

This PR proposes that `advance` abides by what's stated in the
docstring. In practice I was seeing Angstrom report that it consumed
more input than the available input length, which is what led to
`Bigstringaf.blit` (related PR:
https://github.com/inhabitedtype/bigstringaf/pull/25) happening in the
first place.