consume: Add require consume argument to parse functions
For oneshot parse functions, require that the caller specify whether the
parser should consume all of the provided input or just a prefix.

Most parsers are written so that they can consume the prefix of a
string. This combined with the use of backtracking can cause unexpected
parse results where only part of the input was consumed. This is one of
many reasons to avoid backtracking.

In any case, making the behavior explicit in the call will hopefully
demistify some of these results.